completed main HLS pragmas

+++ VivadoHLS/source/AESfunctions.h
@@ -4,11 +4,14 @@
 using namespace std;
 #include <cstdio>
 
-const unsigned int Nb = 4; // columns
-const unsigned int rows = 4; // rows
-const unsigned int stt_lng = Nb * rows; // state length
+// These are AES constants for AES 128, 192, 256
+const unsigned short Nb = 4; // columns (can be changed to a larger number in the future)
+const unsigned short rows = 4; // rows
+const unsigned short stt_lng = Nb * rows; // state length
 
-const unsigned int Nk_max = 8; // max [32-bit words] columns in cipher key
-const unsigned int CipherKeyLenghth_max = Nk_max * rows; // max bytes in key length
-const unsigned int Nr_max = (Nk_max > Nb) ? Nk_max + 6 : Nb + 6; // max number of rounds
-const unsigned int ExtdCipherKeyLenghth_max = (Nr_max + 1) * stt_lng; // max bytes in extended key length
+// NK_max = 8 for AES 256
+// This will allocate (hardware and memory) resources to accommodate  all AES
+const unsigned short Nk_max = 8; // max [32-bit words] columns in cipher key
+const unsigned short CipherKeyLenghth_max = Nk_max * rows; // max bytes in key length
+const unsigned short Nr_max = (Nk_max > Nb) ? Nk_max + 6 : Nb + 6; // max number of rounds
+const unsigned short ExtdCipherKeyLenghth_max = (Nr_max + 1) * stt_lng; // max bytes in extended key length
